Use correct indexes for deopt usage chain lookups
We sometimes have synthetic deopt points. When we have this, the index
present in the deopt usage chain will be the original one that was
cloned to make a synthetic one, and then the index itself is the one
that will be the deopt point index to search for when we are really in
the process of deoptimizing. Handle these properly when building the
deopt materialization table. Fixes one of the crashes caused by the PEA
deopt support in spectest.