Distribute deopt usages when inlining
If we have an invoke, then there might be some registers that have a
deopt usage at that point. This implies that if we deopt, then we must
ensure that the values in those registers are available. Clasically this
has meant that we simply could not delete an instruction that writes to
the register. However, in our PEA new world, we actually *do* delete
such writes, as part of scalar replacement. We are able to do this
thanks to being able to materialize replaced objects as a part of the
deoptimization process.

However, for PEA to correctly analyze this situation, it needs to know
what to materialize at each and every deopt point. If we inline, then at
the point where there once stood an invoke, there now may be many deopt
instructions.

This change calculates the SSA registers that are needed for deopt at
the invoke, by finding which registers have the deopt point listed in
their deopt users. It then, for each deopt point inside of the inline,
places that deopt point on each SSA register that was needed at the
point of the invoke. This means that they now carry a complete set of
deopt points, including the inlinees.

For this to be possible, it was also necesary to stop ignoring inlines
when performing the useless deopt point elimination. Otherwise, all that
we added during the graph merge would be thrown out again, rendering the
whole process useless.